voice-agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-genai==1.16.0
twilio==9.3.5
python-dotenv==1.0.1
websockets==13.1
jinja2==3.1.4
python-multipart==0.0.9
